+++ changelog.md
@@ -1,6 +1,7 @@
 ## 14.04.2023, Fri
 
 - 18:00 - Backend: added Sentry integration.
+- 18:30 - Frontend: tags can be added and removed in Bookmark editor (only existing tags).
 
 ## 13.04.2023, Thu